# cs493/assignment-2-sherburj/public/openapi.yaml
openapi: 3.0.1
info:
  version: 1.0.0
  title: Businesses API
  description: A simple API for a Yelp-like application.
paths:
...
  /photos:
    post:
      summary: Add a new photo
      description: >
        Creates a new photo with specified data and adds it to the
        application's database.
      operationId: addNewPhoto
      tags:
        - photos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Photo'
      responses:
        '201':
          description: New photo successfully added
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: Unique ID of the created photo.
                    example: 10
        '400':
          description: The request body was either not present or did not contain all of the required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/photos/{photoID}':
    parameters:
      - name: photoID
        in: path
        description: Unique ID of a photo.
        schema:
          type: integer
        example: 0
        required: true
    get:
      summary: Fetch data for a specific photo.
      description: >
        Returns complete data for a the photo specified by `photoID`.
      operationId: getPhoto
      tags:
        - photos
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Photo'
        '404':
          description: Specified `photoID` not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update data for a specific photo.
      description: >
        Replaces the data for the photo specified by `photo`.
      operationId: replacePhoto
      tags:
        - photos
      requestBody:
        description: >
          Photo data to replace data for the photo specified by `photoID`.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Photo'
      responses:
        '200':
          description: Success
        '400':
          description: >
            The request body was either not present or did not contain all of the required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: >
            The request tried to change the `userid` or the `businessid` for a photo.  Once a photo is created, these cannot be changed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Specified `photoID` not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Remove a specific photo from the database.
      description: >
        Completely removes the data for the photo specified by `photoID`.
      operationId: removePhoto
      tags:
        - photos
      responses:
        '204':
          description: Success
        '404':
          description: Specified `photoID` not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /users/{userid}/businesses:
    parameters:
      - name: userid
        in: path
        description: Unique ID of a user.
        schema:
          type: integer
        example: 0
        required: true
    get:
      summary: Fetch a user's owned businesses.
      description: >
        Returns a list of all of the businesses owned by the user specified by `userid`.
      operationId: getUserBusinesses
      tags:
        - users
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  lodgings:
                    type: array
                    description: The businesses owned by the specified user.
                    items:
                      $ref: '#/components/schemas/Business'
  /users/{userid}/reviews:
    parameters:
      - name: userid
        in: path
        description: Unique ID of a user.
        schema:
          type: integer
        example: 21
        required: true
    get:
      summary: Fetch a user's reviews.
      description: >
        Returns a list of all of the reviews written by the user specified by `userid`.
      operationId: getUserReviews
      tags:
        - users
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  lodgings:
                    type: array
                    description: The reviews written by the specified user.
                    items:
                      $ref: '#/components/schemas/Review'
  /users/{userid}/photos:
    parameters:
      - name: userid
        in: path
        description: Unique ID of a user.
        schema:
          type: integer
        example: 21
        required: true
    get:
      summary: Fetch a user's photos.
      description: >
        Returns a list of all of the photos posted by the user specified by `userid`.
      operationId: getUserPhotos
      tags:
        - users
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  lodgings:
                    type: array
                    description: The photos posted by the specified user.
                    items:
                      $ref: '#/components/schemas/Photo'
components:
  schemas:
    Business:
      description: >
        An object representing information about a single business.
      type: object
      properties:
        ownerid:
          type: integer
          description: >
            The ID of the user who owns this business.  This is required when creating or updating a business.
          example: 16
        name:
          type: string
          description: Name of the business.
          example: American Dream Pizza
        address:
          type: string
          description: Street address of the business.
          example: 2525 NW Monroe Ave.
        city:
          type: string
          description: The city in which the business resides.
          example: Corvallis
        state:
          type: string
          description: The state in which the business resides.
          example: OR
        zip:
          type: string
          description: The zip code of the city in which the business resides.
          example: '97330'
        phone:
          type: string
          description: The business' phone number.
          example: '541-757-1713'
        category:
          type: string
          description: The business' category.
          example: 'Restaurant'
        subcategory:
          type: string
          description: The business' subcategory.
          example: 'Pizza'
        website:
          type: string
          description: The business' website URL.
          example: 'http://adpizza.com'
        email:
          type: string
          description: The business' email address.
          example: 'brian@adpizza.com'
        reviews:
          type: array
          description: >
            An array of reviews of the business.  These are only returned when fetching detailed information about a business.
          items:
            $ref: '#/components/schemas/Review'
        photos:
          type: array
          description: >
            An array of photos of the business.  These are only returned when fetching detailed information about a business.
          items:
            $ref: '#/components/schemas/Photo'
      required:
        - ownerid
        - name
        - address
        - city
        - state
        - zip
        - phone
        - category
        - subcategory
    Photo:
      description: >
        An object representing information about a photo of a business.
      type: object
      properties:
        userid:
          type: integer
          description: >
            The ID of the user associated with this photo.  This is required when creating or updating a photo.
          example: 7
        businessid:
          type: integer
          description: >
            The ID of the business associated with this photo.  This is required when creating or updating a photo.
          example: 3
        caption:
          type: string
          description: Caption text for the photo
          example: "Kale!"
      required:
        - userid
        - businessid
    Review:
      description: >
        An object representing information about a review of a business.
      type: object
      properties:
        userid:
          type: integer
          description: >
            The ID of the user associated with this review.  This is required when creating or updating a review.
          example: 7
        businessid:
          type: integer
          description: >
            The ID of the business associated with this review.  This is required when creating or updating a review.
          example: 3
        dollars:
          type: integer
          description: A rating of the expensiveness of a business.  Must be an integer between 1 and 4.
          example: 3
          minimum: 1
          maximum: 4
        stars:
          type: number
          description: A rating of the quality of a business.  Must be an number between 0 and 5.
          example: 5
          minimum: 0
          maximum: 5
        review:
          type: string
          description: Textual business review.
          example: "Great produce!"
      required:
        - userid
        - businessid
        - dollars
        - stars
    Error:
      description: >
        An object representing an error response from the API.
      type: object
      properties:
        error:
          type: string
          description: A message describing the error.
...